move terminate_cluster into utils
